gsqlite3: Ensure module is not loaded when invoked directly
